+++ cms/src/collections/Authors.ts
@@ -5,7 +5,7 @@ import {
   triggerDeployAfterChange,
   triggerDeployAfterDelete,
 } from '../hooks/triggerDeploy'
-import { AUTOSAVE_INTERVAL, livePreviewBreakpoints } from '../shared'
+import { draftVersions, livePreviewBreakpoints } from '../shared'
 
 export const Authors: PageCollectionConfig = {
   slug: 'authors',
@@ -14,12 +14,7 @@ export const Authors: PageCollectionConfig = {
     defaultColumns: ['name', 'path', 'updatedAt', '_status'],
     livePreview: { breakpoints: [...livePreviewBreakpoints] },
   },
-  versions: {
-    drafts: {
-      autosave: { interval: AUTOSAVE_INTERVAL },
-      schedulePublish: false,
-    },
-  },
+  versions: draftVersions,
   hooks: {
     afterChange: [triggerDeployAfterChange],
     afterDelete: [triggerDeployAfterDelete],